+++ secure-minify/js/serverlist.js
@@ -45,8 +45,9 @@ document.getElementById('saveNewButton').removeAttribute('hidden');document.getE
 ;document.getElementById('proxyCheckboxId').checked=true;document.getElementById('autoReconnectCheckboxId').checked=false;document.getElementById('loggingCheckboxId').checked=false
 ;document.getElementById('passwordInputId').setAttribute('disabled','');document.getElementById('passwordInputId').value='(Hidden)';document.getElementById('identifyNickInputId').value=''
 ;document.getElementById('identifyCommandInputId').setAttribute('disabled','');document.getElementById('identifyCommandInputId').value='(Hidden)';document.getElementById('nickInputId').value=''
-;document.getElementById('userInputId').value='';document.getElementById('realInputId').value='';document.getElementById('modesInputId').value='';document.getElementById('channelListInputId').value=''
-;resolve(null)});const populateIrcServerForm=data=>new Promise((resolve,reject)=>{clearIrcServerForm();document.getElementById('saveNewButton').setAttribute('hidden','')
+;document.getElementById('altNickInputId').value='';document.getElementById('recoverNickCheckboxId').checked=false;document.getElementById('userInputId').value=''
+;document.getElementById('realInputId').value='';document.getElementById('modesInputId').value='';document.getElementById('channelListInputId').value='';resolve(null)})
+;const populateIrcServerForm=data=>new Promise((resolve,reject)=>{clearIrcServerForm();document.getElementById('saveNewButton').setAttribute('hidden','')
 ;document.getElementById('saveModifiedButton').removeAttribute('hidden');document.getElementById('serverListDisabledDiv').setAttribute('hidden','')
 ;document.getElementById('warningVisibilityDiv').setAttribute('hidden','');document.getElementById('listVisibilityDiv').setAttribute('hidden','')
 ;document.getElementById('formVisibilityDiv').removeAttribute('hidden');document.getElementById('serverPasswordWarningDiv').setAttribute('hidden','')
@@ -61,12 +62,13 @@ document.getElementById('saveNewButton').removeAttribute('hidden');document.getE
 ;document.getElementById('passwordInputId').setAttribute('disabled','');document.getElementById('passwordInputId').value='(hidden)'
 ;document.getElementById('identifyNickInputId').value=data.identifyNick;document.getElementById('identifyCommandInputId').setAttribute('disabled','')
 ;document.getElementById('identifyCommandInputId').value='(hidden)';document.getElementById('serverPasswordWarningDiv').setAttribute('hidden','')
-;document.getElementById('nickservCommandWarningDiv').setAttribute('hidden','');document.getElementById('nickInputId').value=data.nick;document.getElementById('userInputId').value=data.user
-;document.getElementById('realInputId').value=data.real;document.getElementById('modesInputId').value=data.modes;document.getElementById('channelListInputId').value=data.channelList;resolve(null)})
-;const openIrcServerEdit=index=>{_clearError();clearIrcServerForm().then(()=>fetchServerList(index,1)).then(data=>populateIrcServerForm(data)).catch(err=>{_showError(err.toString()||err)
-;console.log(err)})};const copyIrcServerToNew=index=>{_clearError();submitServer({index:index,action:'duplicate'
-},'COPY',index).then(data=>checkErrorAndCloseEdit(data)).then(()=>fetchServerList(-1,-1)).then(data=>buildServerListTable(data)).catch(err=>{_showError(err.toString()||err);console.log(err)})}
-;const deleteIrcServer=index=>{_clearError();submitServer({index:index
+;document.getElementById('nickservCommandWarningDiv').setAttribute('hidden','');document.getElementById('nickInputId').value=data.nick;document.getElementById('altNickInputId').value=data.altNick
+;if(0===data.altNick.length)document.getElementById('recoverNickCheckboxId').checked=false;else if(data.recoverNick)document.getElementById('recoverNickCheckboxId').checked=true;else document.getElementById('recoverNickCheckboxId').checked=false
+;document.getElementById('userInputId').value=data.user;document.getElementById('realInputId').value=data.real;document.getElementById('modesInputId').value=data.modes
+;document.getElementById('channelListInputId').value=data.channelList;resolve(null)});const openIrcServerEdit=index=>{_clearError()
+;clearIrcServerForm().then(()=>fetchServerList(index,1)).then(data=>populateIrcServerForm(data)).catch(err=>{_showError(err.toString()||err);console.log(err)})};const copyIrcServerToNew=index=>{
+_clearError();submitServer({index:index,action:'duplicate'},'COPY',index).then(data=>checkErrorAndCloseEdit(data)).then(()=>fetchServerList(-1,-1)).then(data=>buildServerListTable(data)).catch(err=>{
+_showError(err.toString()||err);console.log(err)})};const deleteIrcServer=index=>{_clearError();submitServer({index:index
 },'DELETE',index).then(data=>checkErrorAndCloseEdit(data)).then(()=>fetchServerList(-1,-1)).then(data=>buildServerListTable(data)).catch(err=>{_showError(err.toString()||err);console.log(err)})}
 ;const moveUpInList=index=>{_clearError();submitServer({index:index,action:'move-up'
 },'COPY',index).then(data=>checkErrorAndCloseEdit(data)).then(()=>fetchServerList(-1,-1)).then(data=>buildServerListTable(data)).catch(err=>{_showError(err.toString()||err);console.log(err)})}
@@ -79,11 +81,13 @@ document.getElementById('saveNewButton').removeAttribute('hidden');document.getE
 ;if(!document.getElementById('passwordInputId').hasAttribute('disabled'))data.password=document.getElementById('passwordInputId').value
 ;data.identifyNick=document.getElementById('identifyNickInputId').value
 ;if(!document.getElementById('identifyCommandInputId').hasAttribute('disabled'))data.identifyCommand=document.getElementById('identifyCommandInputId').value
-;data.nick=document.getElementById('nickInputId').value;data.user=document.getElementById('userInputId').value;data.real=document.getElementById('realInputId').value
-;data.modes=document.getElementById('modesInputId').value;data.channelList=document.getElementById('channelListInputId').value;let errorStr=null;if(''===data.name)errorStr='Label is required input.'
-;if(''===data.host)errorStr='Host/IP is required input.';if(isNaN(data.port))errorStr='Invalid port number';if(''===data.nick)errorStr='Nickname is required input.'
-;if(''===data.user)errorStr='Unix ident user is required input.';if(''===data.real)errorStr='Real Name is required input.';if(errorStr)reject(new Error(errorStr));else resolve({data:data,index:index})
-});const buildServerListTable=data=>new Promise((resolve,reject)=>{const tableNode=document.getElementById('tbodyId');while(tableNode.firstChild)tableNode.removeChild(tableNode.firstChild)
+;data.nick=document.getElementById('nickInputId').value;data.altNick=document.getElementById('altNickInputId').value
+;if(0===data.altNick.length)data.recoverNick=false;else if(document.getElementById('recoverNickCheckboxId').checked)data.recoverNick=true;else data.recoverNick=false
+;data.user=document.getElementById('userInputId').value;data.real=document.getElementById('realInputId').value;data.modes=document.getElementById('modesInputId').value
+;data.channelList=document.getElementById('channelListInputId').value;let errorStr=null;if(''===data.name)errorStr='Label is required input.';if(''===data.host)errorStr='Host/IP is required input.'
+;if(isNaN(data.port))errorStr='Invalid port number';if(''===data.nick)errorStr='Nickname is required input.';if(''===data.user)errorStr='Unix ident user is required input.'
+;if(''===data.real)errorStr='Real Name is required input.';if(errorStr)reject(new Error(errorStr));else resolve({data:data,index:index})})
+;const buildServerListTable=data=>new Promise((resolve,reject)=>{const tableNode=document.getElementById('tbodyId');while(tableNode.firstChild)tableNode.removeChild(tableNode.firstChild)
 ;const columnTitles=['Index','Disabled','Label','Host','Port','Nick','','','',''];const titleRowEl=document.createElement('tr');columnTitles.forEach(titleName=>{const tdEl=document.createElement('td')
 ;tdEl.textContent=titleName;titleRowEl.appendChild(tdEl)});tableNode.appendChild(titleRowEl);if(Array.isArray(data)&&data.length>0)for(let i=0;i<data.length;i++){
 const rowEl=document.createElement('tr');rowEl.setAttribute('index',i.toString());const td01El=document.createElement('td');const td02El=document.createElement('td')